+++ apps/prairielearn/src/lib/github.js
@@ -520,17 +520,14 @@ module.exports = {
             status: 'failed',
             course_request_id: options.course_request_id,
           });
-          opsbot.sendCourseRequestMessage(
-            `*Failed to create course "${options.short_name}"*\n\n` +
-              '```\n' +
-              `${err.message.trim()}\n` +
-              '```',
-            (err) => {
-              ERR(err, () => {
-                logger.error(err);
-              });
-            },
-          );
+          opsbot
+            .sendCourseRequestMessage(
+              `*Failed to create course "${options.short_name}"*\n\n` +
+                '```\n' +
+                `${err.message.trim()}\n` +
+                '```',
+            )
+            .catch((err) => logger.error(err));
         }
       },
     );