+++ packages/aws-imds/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/aws-imds
 
+## 1.0.8
+
+### Patch Changes
+
+- bd0053577: Upgrade all dependencies
+
+## 1.0.7
+
+### Patch Changes
+
+- f03853d90: Upgrade all dependencies
+
 ## 1.0.6
 
 ### Patch Changes