+++ packages/flash/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/flash
 
+## 1.0.6
+
+### Patch Changes
+
+- bd0053577: Upgrade all dependencies
+
+## 1.0.5
+
+### Patch Changes
+
+- f03853d90: Upgrade all dependencies
+
 ## 1.0.4
 
 ### Patch Changes