+++ packages/html-ejs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prairielearn/html-ejs
 
+## 1.1.10
+
+### Patch Changes
+
+- bd0053577: Upgrade all dependencies
+- Updated dependencies [bd0053577]
+  - @prairielearn/html@3.0.8
+
+## 1.1.9
+
+### Patch Changes
+
+- f03853d90: Upgrade all dependencies
+- Updated dependencies [f03853d90]
+  - @prairielearn/html@3.0.7
+
 ## 1.1.8
 
 ### Patch Changes