+++ packages/named-locks/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prairielearn/named-locks
 
+## 1.5.5
+
+### Patch Changes
+
+- bd0053577: Upgrade all dependencies
+- Updated dependencies [bd0053577]
+  - @prairielearn/postgres@1.7.7
+
+## 1.5.4
+
+### Patch Changes
+
+- f03853d90: Upgrade all dependencies
+- Updated dependencies [f03853d90]
+  - @prairielearn/postgres@1.7.6
+
 ## 1.5.3
 
 ### Patch Changes