+++ packages/sentry/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prairielearn/sentry
 
+## 1.1.1
+
+### Patch Changes
+
+- bd0053577: Upgrade all dependencies
+
+## 1.1.0
+
+### Minor Changes
+
+- 7f31ebd78: Export `runWithAsyncContext` function
+
+### Patch Changes
+
+- f03853d90: Upgrade all dependencies
+
 ## 1.0.8
 
 ### Patch Changes